requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
--index-url https://pypi.tuna.tsinghua.edu.cn/simple

asgiref==3.4.1
    # via django
attrs==21.4.0
    # via pytest
click==8.0.3
    # via pip-tools
django==4.0.1
    # via -r requirements.in
iniconfig==1.1.1
    # via pytest
mysqlclient==2.1.0
    # via -r requirements.in
packaging==21.3
    # via pytest
pep517==0.12.0
    # via pip-tools
pip-tools==6.4.0
    # via -r requirements.in
pluggy==1.0.0
    # via pytest
py==1.11.0
    # via pytest
pybase62==0.4.3
    # via -r requirements.in
pyparsing==3.0.6
    # via packaging
pytest==6.2.5
    # via
    #   -r requirements.in
    #   pytest-django
    #   pytest-mock
pytest-django==4.5.2
    # via -r requirements.in
pytest-mock==3.6.1
    # via -r requirements.in
sqlparse==0.4.2
    # via django
toml==0.10.2
    # via pytest
tomli==2.0.0
    # via pep517
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
